Add GlobalStyle and ThemeProvider
